requirements.txt
checksumdir
python-dotenv